+++ broadcast-extension-sample/src/main/java/ru/surfstudio/android/broadcast/extension/sample/ui/screen/main/MainPresenter.kt
@@ -2,9 +2,12 @@ package ru.surfstudio.android.broadcast.extension.sample.ui.screen.main
 
 import android.telephony.SmsManager
 import ru.surfstudio.android.broadcast.extension.sample.interactor.SmsBroadcastReceiver
+import ru.surfstudio.android.broadcast.extension.sample.interactor.SmsPermissionRequest
 import ru.surfstudio.android.core.mvp.presenter.BasePresenter
 import ru.surfstudio.android.core.mvp.presenter.BasePresenterDependency
+import ru.surfstudio.android.core.ui.permission.PermissionManager
 import ru.surfstudio.android.dagger.scope.PerScreen
+import ru.surfstudio.android.logger.Logger
 import ru.surfstudio.android.message.MessageController
 import javax.inject.Inject
 
@@ -12,12 +15,15 @@ import javax.inject.Inject
  * Презентер главного экрана
  */
 @PerScreen
-internal class MainPresenter @Inject constructor(basePresenterDependency: BasePresenterDependency,
-                                                 private val messageController: MessageController,
-                                                 private val smsBroadcastReceiver: SmsBroadcastReceiver
+internal class MainPresenter @Inject constructor(
+        basePresenterDependency: BasePresenterDependency,
+        private val permissionManager: PermissionManager,
+        private val messageController: MessageController,
+        private val smsBroadcastReceiver: SmsBroadcastReceiver
 ) : BasePresenter<MainActivityView>(basePresenterDependency) {
 
     private val sm: MainScreenModel = MainScreenModel()
+    private val smsPermissionRequest = SmsPermissionRequest()
 
     override fun onFirstLoad() {
         subscribe(smsBroadcastReceiver.observeBroadcast()) { sms ->
@@ -34,13 +40,18 @@ internal class MainPresenter @Inject constructor(basePresenterDependency: BasePr
 
     fun sendSms(phoneNumber: String, message: String) {
         if (phoneNumber.isNotEmpty()) {
-            SmsManager.getDefault()
-                    .sendTextMessage(
-                            phoneNumber,
-                            null,
-                            message,
-                            null,
-                            null)
+            subscribe(permissionManager.request(smsPermissionRequest).toObservable(), {
+                SmsManager.getDefault()
+                        .sendTextMessage(
+                                phoneNumber,
+                                null,
+                                message,
+                                null,
+                                null
+                        )
+            }, {
+                Logger.e(it)
+            })
         }
     }
 }