+++ core-ui/src/main/java/androidx/fragment/app/BackstackAccessor.java
@@ -17,9 +17,6 @@
 
 package androidx.fragment.app;
 
-import androidx.fragment.app.Fragment;
-import androidx.fragment.app.FragmentManager;
-
 /**
  * With this utility class one can access the fragment backstack
  *