+++ easyadapter-sample/src/main/java/ru/surfstudio/android/easyadapter/sample/ui/screen/multitype/MultitypeListActivityView.kt
@@ -10,8 +10,6 @@ import ru.surfstudio.android.core.mvp.presenter.Presenter
 import ru.surfstudio.android.easyadapter.EasyAdapter
 import ru.surfstudio.android.easyadapter.ItemList
 import ru.surfstudio.android.easyadapter.sample.R
-import ru.surfstudio.android.easyadapter.sample.domain.FirstData
-import ru.surfstudio.android.easyadapter.sample.domain.SecondData
 import ru.surfstudio.android.easyadapter.sample.ui.base.configurator.CustomActivityScreenConfigurator
 import ru.surfstudio.android.easyadapter.sample.ui.screen.common.controllers.EmptyItemController
 import ru.surfstudio.android.easyadapter.sample.ui.screen.common.controllers.FirstDataItemController
@@ -29,19 +27,9 @@ class MultitypeListActivityView : BaseRenderableActivityView<MultitypeListScreen
     private val emptyItemController = EmptyItemController()
     private val twoDataItemController = TwoDataItemController()
 
-    private val firstDataItemController = FirstDataItemController(
-            object : FirstDataItemController.FirstDataClickListener {
-                override fun onClick(firstData: FirstData) {
-                    toast("Value = $firstData")
-                }
-            })
+    private val firstDataItemController = FirstDataItemController { toast("Value = $it") }
 
-    private val secondDataItemController = SecondDataItemController(
-            object : SecondDataItemController.SecondDataClickListener {
-                override fun onClick(secondData: SecondData) {
-                    toast("Value = ${secondData.stringValue}")
-                }
-            })
+    private val secondDataItemController = SecondDataItemController { toast("Value = ${it.stringValue}") }
 
     override fun createConfigurator(): CustomActivityScreenConfigurator {
         return MultitypeListScreenConfigurator(intent)
@@ -62,10 +50,10 @@ class MultitypeListActivityView : BaseRenderableActivityView<MultitypeListScreen
 
     override fun renderInternal(sm: MultitypeListScreenModel) {
         adapter.setItems(ItemList.create()
-                .add(emptyItemController)
-                .add(sm.firstData, sm.secondData, twoDataItemController)
                 .addAll(sm.firstDataList, firstDataItemController)
-                .addAll(sm.secondDataList, secondDataItemController))
+                .addAll(sm.secondDataList, secondDataItemController)
+                .add(emptyItemController)
+                .add(sm.firstData, sm.secondData, twoDataItemController))
     }
 
     private fun initRecycler() {