+++ template/app-injector/src/main/java/ru/surfstudio/standard/app_injector/ui/screen/configurator/storage/ScreenConfiguratorStorage.kt
@@ -3,14 +3,16 @@ package ru.surfstudio.standard.app_injector.ui.screen.configurator.storage
 import android.content.Intent
 import android.os.Bundle
 import ru.surfstudio.standard.app_injector.ui.configurator.*
-import ru.surfstudio.standard.app_injector.ui.screen.configurator.activity.debug.FcmDebugScreenConfigurator
-import ru.surfstudio.standard.app_injector.ui.screen.configurator.activity.debug.DebugScreenConfigurator
 import ru.surfstudio.standard.app_injector.ui.screen.configurator.activity.MainScreenConfigurator
 import ru.surfstudio.standard.app_injector.ui.screen.configurator.activity.SplashScreenConfigurator
-import ru.surfstudio.standard.app_injector.ui.screen.configurator.activity.debug.CommonControllersDebugScreenConfigurator
+import ru.surfstudio.standard.app_injector.ui.screen.configurator.activity.debug.*
 import ru.surfstudio.standard.f_debug.debug.DebugActivityView
 import ru.surfstudio.standard.f_debug.common_controllers.CommonControllersDebugActivityView
 import ru.surfstudio.standard.f_debug.fcm.FcmDebugActivityView
+import ru.surfstudio.standard.f_debug.info.AppInfoDebugActivityView
+import ru.surfstudio.standard.f_debug.memory.MemoryDebugActivityView
+import ru.surfstudio.standard.f_debug.server_settings.ServerSettingsDebugActivityView
+import ru.surfstudio.standard.f_debug.server_settings.reboot.RebootActivityDebugView
 import ru.surfstudio.standard.f_main.MainActivityView
 import ru.surfstudio.standard.f_splash.SplashActivityView
 import kotlin.reflect.KClass
@@ -24,6 +26,10 @@ object ScreenConfiguratorStorage {
                 put(DebugActivityView::class) { DebugScreenConfigurator(it) }
                 put(FcmDebugActivityView::class) { FcmDebugScreenConfigurator(it) }
                 put(CommonControllersDebugActivityView::class) { CommonControllersDebugScreenConfigurator(it) }
+                put(ServerSettingsDebugActivityView::class) { ServerSettingsDebugScreenConfigurator(it) }
+                put(RebootActivityDebugView::class) { RebootDebugScreenConfigurator(it) }
+                put(MemoryDebugActivityView::class) { MemoryDebugScreenConfigurator(it) }
+                put(AppInfoDebugActivityView::class) { AppInfoDebugScreenConfigurator(it) }
             }
 
     val activityConfiguratorMap = HashMap<KClass<*>, (intent: Intent) -> ActivityConfigurator>()
@@ -35,7 +41,7 @@ object ScreenConfiguratorStorage {
 
             }
 
-    val dialogScreenConfiguratorMap= HashMap<KClass<*>, (args: Bundle) -> DialogScreenConfigurator>()
+    val dialogScreenConfiguratorMap = HashMap<KClass<*>, (args: Bundle) -> DialogScreenConfigurator>()
             .apply {
 
             }