+++ template/f-debug/src/main/java/ru/surfstudio/standard/f_debug/debug/DebugPresenter.kt
@@ -6,6 +6,9 @@ import ru.surfstudio.android.core.ui.navigation.activity.navigator.ActivityNavig
 import ru.surfstudio.android.dagger.scope.PerScreen
 import ru.surfstudio.standard.f_debug.common_controllers.CommonControllersDebugActivityRoute
 import ru.surfstudio.standard.f_debug.fcm.FcmDebugActivityRoute
+import ru.surfstudio.standard.f_debug.info.AppInfoDebugActivityRoute
+import ru.surfstudio.standard.f_debug.memory.MemoryDebugActivityRoute
+import ru.surfstudio.standard.f_debug.server_settings.ServerSettingsDebugActivityRoute
 import javax.inject.Inject
 
 /**
@@ -17,11 +20,23 @@ class DebugPresenter @Inject constructor(
         private val activityNavigator: ActivityNavigator
 ) : BasePresenter<DebugActivityView>(basePresenterDependency) {
 
+    fun openServerSettingsScreen() {
+        activityNavigator.start(ServerSettingsDebugActivityRoute())
+    }
+
     fun openControllersScreen() {
         activityNavigator.start(CommonControllersDebugActivityRoute())
     }
 
     fun openFcmTokenScreen() {
         activityNavigator.start(FcmDebugActivityRoute())
+    }
+
+    fun openMemoryScreen() {
+        activityNavigator.start(MemoryDebugActivityRoute())
+    }
+
+    fun openAppInfoScreen() {
+        activityNavigator.start(AppInfoDebugActivityRoute())
     }
 }