+++ template/f-debug/src/main/java/ru/surfstudio/standard/f_debug/info/AppInfoDebugActivityView.kt
@@ -1,0 +1,49 @@
+package ru.surfstudio.standard.f_debug.info
+
+import android.annotation.SuppressLint
+import android.os.Bundle
+import android.os.PersistableBundle
+import androidx.annotation.LayoutRes
+import kotlinx.android.synthetic.main.activity_app_info_debug.*
+import ru.surfstudio.android.core.mvp.activity.BaseRenderableActivityView
+import ru.surfstudio.android.core.mvp.presenter.CorePresenter
+import ru.surfstudio.android.template.f_debug.BuildConfig
+import ru.surfstudio.android.template.f_debug.R
+import ru.surfstudio.standard.base_ui.provider.component.ComponentProvider
+import javax.inject.Inject
+
+/**
+ * Вью экрана показа общей информации
+ */
+class AppInfoDebugActivityView : BaseRenderableActivityView<AppInfoDebugScreenModel>() {
+
+    @Inject
+    lateinit var presenter: AppInfoDebugPresenter
+
+    override fun getPresenters(): Array<CorePresenter<*>> = arrayOf(presenter)
+
+    override fun createConfigurator() = ComponentProvider.createActivityScreenConfigurator(intent, this::class)
+
+    @LayoutRes
+    override fun getContentView(): Int = R.layout.activity_app_info_debug
+
+    override fun getScreenName(): String = "debug_info"
+
+    override fun onCreate(savedInstanceState: Bundle?, persistentState: PersistableBundle?, viewRecreated: Boolean) {
+        super.onCreate(savedInstanceState, persistentState, viewRecreated)
+
+        initContent()
+    }
+
+    override fun renderInternal(sm: AppInfoDebugScreenModel) {}
+
+    @SuppressLint("StringFormatMatches")
+    private fun initContent() {
+        with(packageManager.getPackageInfo(packageName, 0)) {
+            app_info_version_code_tv.text = getString(R.string.app_info_version_code_text, longVersionCode)
+            app_info_version_name_tv.text = getString(R.string.app_info_version_name_text, versionName)
+        }
+        app_info_package_name_tv.text = getString(R.string.app_info_package_name_text, packageName)
+        app_info_build_type_tv.text = getString(R.string.app_info_build_type_text, BuildConfig.BUILD_TYPE)
+    }
+}